+++ tests/driver.test.ts
@@ -1,6 +1,6 @@
 import { Okareo } from '../dist';
 import { getProjectId } from './setup-env';
-import { OpenAIModel, TestRunType, MultiTurnDriver } from "../dist";
+import { OpenAIModel, TestRunType, MultiTurnDriver, DriverParameters, ModelInvocation, RunTestProps, CustomMultiturnTarget } from "../dist";
 
 const OKAREO_API_KEY = process.env.OKAREO_API_KEY || "<YOUR_OKAREO_KEY>";
 const UNIQUE_BUILD_ID = (process.env.SDK_BUILD_ID || `local.${(Math.random() + 1).toString(36).substring(7)}`);
@@ -14,7 +14,6 @@ const SEED_DATA = [
         "result": "hello world"
     },
 ];
-
 
 describe('Drivers', () => {
     beforeAll(async () => {
@@ -36,7 +35,10 @@ describe('Drivers', () => {
             project_id: project_id,
             models: {
                 type: "driver",
-                driver_params: {"driver_type": "openai"},
+                driver_params: {
+                    driver_temperature: 0,
+                    repeats: 1,
+                } as DriverParameters,
                 target: {
                     type: "openai",
                     model_id: "gpt-4o-mini",
@@ -60,4 +62,75 @@ describe('Drivers', () => {
         expect(data.model_metrics).toBeDefined();
     });
 
+    test('Custom Multiturn Evaluation', async () => {
+        const okareo = new Okareo({ api_key: OKAREO_API_KEY });
+
+        const project_id = (await okareo.getProjects()).find(p => p.name === 'Global')?.id;
+
+        const prompt_template = (text: string) => `You are interacting with an agent who is good at answering questions.\n\nAsk them a very simple math question. ${text} insist that they answer the question, even if they try to avoid it.`;
+
+        const off_topic_directive = 'You should only engage in conversation about WebBizz, the e-commerce platform.';
+
+        const scenario: any = await okareo.create_scenario_set({
+            name: `Multi-turn Demo Scenario - ${(Math.random() + 1).toString(36).substring(7)}`,
+            project_id: project_id || '',
+            seed_data: [
+                { input: prompt_template("Rudely"), result: off_topic_directive },
+                { input: prompt_template("Politely"), result: off_topic_directive }]
+        });
+
+        const polite_chatbot = {
+            type: 'custom_target',
+            invoke: async (input_: any[]) => {
+                const common_response_args = {
+                    model_input: input_,
+                    model_output_metadata: {},
+                }
+                if (input_.length < 2) {
+                    return {
+                        model_prediction: "Hi! I'm a chatbot that can help you with WebBizz, an e-commerce platform. Ask me anything about WebBizz!",
+                        ...common_response_args
+                    } as ModelInvocation;
+                }
+                const message_data = input_[input_.length - 1];
+                const user_message: string = message_data.content;
+                let response: string;
+                if (user_message.toLowerCase().includes("please")) {
+                    response = "Yes, I'm happy to do whatever you'd like me to do!";
+                } else {
+                    response = "I'm only here to talk about WebBizz. How can I help you with that?";
+                }
+                return {
+                    model_prediction: response,
+                    ...common_response_args
+                } as ModelInvocation;
+            }
+        } as CustomMultiturnTarget;
+
+        const model = await okareo.register_model({
+            name: 'Demo MultiTurnDrivera',
+            project_id: project_id || '',
+            models: {
+                type: 'driver',
+                driver_params: {
+                    driver_temperature: 1,
+                    max_turns: 2,
+                    repeats: 1,
+                } as DriverParameters,
+                target: polite_chatbot
+            } as MultiTurnDriver,
+            update: true,
+        });
+
+        const data: any = await model.run_test({
+            name: 'Multi-turn Demo Evaluation',
+            project_id: project_id,
+            scenario_id: scenario.scenario_id,
+            calculate_metrics: true,
+            type: TestRunType.NL_GENERATION,
+            checks: ['behavior_adherence'],
+        } as RunTestProps);
+        expect(data).toBeDefined();
+        expect(data.model_metrics).toBeDefined();
+    });
 });
